requirements.txt
torch
torchvision
torchaudio
sentence-transformers
faiss-cpu
llama-cpp-python
streamlit
